task 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -16,5 +16,6 @@ if __name__ == '__main__':
     sess = sessionmaker(bind=engine)()
 
     states = sess.query(State).filter(State.name.contains('a'))
-    states.delete()
-    sess.commit()
+    if states:
+        states.delete()
+        sess.commit()
